:recycle: (901): extract pure transforms from god-object into domain/transcription
