+++ mcr-core/mcr_meeting/app/domain/transcription/chunking.py
@@ -3,8 +3,7 @@
 from collections.abc import Iterable
 
 from mcr_meeting.app.configs.base import WhisperTranscriptionSettings
-from mcr_meeting.app.schemas.transcription_schema import DiarizationSegment
-from mcr_meeting.app.services.speech_to_text.utils.types import TimeSpan
+from mcr_meeting.app.schemas.transcription_schema import DiarizationSegment, TimeSpan
 
 _settings = WhisperTranscriptionSettings()
 MAX_CHUNK_DURATION = _settings.MAX_CHUNK_DURATION
@@ -14,7 +13,6 @@ SPLIT_SEARCH_WINDOW_RATIO = _settings.SPLIT_SEARCH_WINDOW_RATIO
 def _merge_overlapping_intervals(
     spans: Iterable[TimeSpan],
 ) -> list[TimeSpan]:
-    """Merge overlapping/touching TimeSpans into non-overlapping intervals."""
     sorted_spans = sorted(spans, key=lambda s: s.start)
     if not sorted_spans:
         return []
@@ -37,18 +35,11 @@ def _find_split_boundary(
     max_chunk_duration: float,
     split_search_window_ratio: float = SPLIT_SEARCH_WINDOW_RATIO,
 ) -> float:
-    """Find the best boundary to split a chunk that would exceed max duration.
-
-    Looks for the largest silence gap in the last ``split_search_window_ratio``
-    of the chunk. Falls back to the last available gap, or an arbitrary cut at
-    max_chunk_duration.
-    """
     window_start = chunk_start + (1 - split_search_window_ratio) * max_chunk_duration
     window_end = chunk_start + max_chunk_duration
 
     all_intervals = chunk_intervals + [next_interval]
 
-    # Collect all silences (gaps between consecutive intervals)
     best_gap: TimeSpan | None = None
 
     for j in range(len(all_intervals) - 1):
@@ -64,7 +55,6 @@ def _find_split_boundary(
     if best_gap is not None:
         return best_gap.midpoint
 
-    # No gap at all — hard cut
     return window_end
 
 
@@ -72,20 +62,6 @@ def compute_transcription_chunks(
     diarization: list[DiarizationSegment],
     max_chunk_duration: float = MAX_CHUNK_DURATION,
 ) -> list[TimeSpan]:
-    """Compute large transcription chunks from diarization segments.
-
-    Merges overlapping diarization segments, then greedily accumulates them
-    into chunks up to ``max_chunk_duration``. When a chunk would exceed the
-    limit, the boundary is placed at the midpoint of the largest silence in
-    the last portion (controlled by ``split_search_window_ratio``) of the chunk.
-
-    Args:
-        diarization: Diarization segments (may overlap).
-        max_chunk_duration: Maximum chunk length in seconds (default 600).
-
-    Returns:
-        List of TimeSpan chunks covering all speech regions.
-    """
     if not diarization:
         return []
 
@@ -95,10 +71,8 @@ def compute_transcription_chunks(
     if not merged:
         return []
 
-    # Greedy accumulation
     chunks: list[TimeSpan] = []
     chunk_start = merged[0].start
-    # Track the intervals belonging to the current chunk
     chunk_intervals: list[TimeSpan] = [merged[0]]
 
     for i in range(1, len(merged)):
@@ -110,7 +84,6 @@ def compute_transcription_chunks(
             chunk_intervals.append(interval)
             continue
 
-        # Need to split: find the best silence in the search window
         boundary = _find_split_boundary(
             chunk_start, chunk_intervals, interval, max_chunk_duration
         )
@@ -119,7 +92,6 @@ def compute_transcription_chunks(
         chunk_start = boundary
         chunk_intervals = [interval]
 
-    # Final chunk
     last_end = chunk_intervals[-1].end
     chunks.append(TimeSpan(chunk_start, last_end))
