+++ mcr-core/mcr_meeting/app/schemas/transcription_schema.py
@@ -1,9 +1,48 @@
+from dataclasses import dataclass
 from enum import StrEnum
 from io import BytesIO
 
 from pydantic import BaseModel, Field, field_validator
 
 from mcr_meeting.app.exceptions.exceptions import UnknownDiarizationStatus
+
+
+@dataclass
+class TimeSpan:
+    """Small helper to reason about time intervals in seconds."""
+
+    start: float
+    end: float
+
+    def __post_init__(self) -> None:
+        if self.end < self.start:
+            raise ValueError(
+                f"Invalid TimeSpan with end < start ({self.start} > {self.end})"
+            )
+
+    @property
+    def duration(self) -> float:
+        return self.end - self.start
+
+    @property
+    def midpoint(self) -> float:
+        return (self.start + self.end) / 2.0
+
+    def touches_or_overlaps(self, other: "TimeSpan") -> bool:
+        return not (self.end < other.start or self.start > other.end)
+
+    def overlap(self, other: "TimeSpan") -> float:
+        start = max(self.start, other.start)
+        end = min(self.end, other.end)
+        return max(0.0, end - start)
+
+    def merge(self, other: "TimeSpan") -> "TimeSpan":
+        return TimeSpan(min(self.start, other.start), max(self.end, other.end))
+
+    def gap_to(self, other: "TimeSpan") -> "TimeSpan | None":
+        if self.end >= other.start:
+            return None
+        return TimeSpan(self.end, other.start)
 
 
 class SpeakerTranscription(BaseModel):
@@ -50,6 +89,30 @@ class DiarizationSegment(BaseModel):
     start: float
     end: float
     speaker: str
+
+
+class Participant(BaseModel):
+    speaker_id: str = Field(
+        description="Identifiant unique du locuteur dans la transcription ex: LOCUTEUR_03.",
+    )
+    name: str | None = Field(
+        None,
+        description="Prenom et/ou nom déduit pour le locuteur à partir des interactions dans la transcription. Ex: 'Jean' ou 'Jean Dupont'.",
+    )
+    role: str | None = Field(
+        None,
+        description="Fonction/rôle si mentionné ou déduit (ex. PO, Tech Lead, Directeur financier).",
+    )
+    confidence: float | None = Field(
+        ge=0.0,
+        le=1.0,
+        description="Niveau de confiance (entre 0 et 1) indiquant à quel point tu es certain du nom associé locuteur.",
+    )
+    association_justification: str | None = Field(
+        description=(
+            "Identification explicite ou déduction par contexte ayant permis d'associer ce nom/rôle au locuteur avec l'id."
+        ),
+    )
 
 
 class DiarizationJobStatus(StrEnum):
