+++ mcr-core/mcr_meeting/app/services/speech_to_text/transcription_processor.py
@@ -16,6 +16,7 @@ from mcr_meeting.app.infrastructure.unleash import (
     get_feature_flag_client,
 )
 from mcr_meeting.app.schemas.transcription_schema import (
+    TimeSpan,
     TranscriptionSegment,
 )
 from mcr_meeting.app.services.speech_to_text.utils.audio import (
@@ -24,7 +25,6 @@ from mcr_meeting.app.services.speech_to_text.utils.audio import (
 from mcr_meeting.app.services.speech_to_text.utils.models import (
     get_transcription_model,
 )
-from mcr_meeting.app.services.speech_to_text.utils.types import TimeSpan
 
 transcription_settings = WhisperTranscriptionSettings()
 api_settings = TranscriptionApiSettings()
