+++ mcr-core/tests/unit/test_audio_chunk_download.py
@@ -13,9 +13,7 @@ from mcr_meeting.app.schemas.S3_types import S3Object
 class TestDownloadAndConcatenateS3AudioChunks:
     """Tests for the function download_and_concatenate_s3_audio_chunks_into_bytes."""
 
-    @patch(
-        "mcr_meeting.app.infrastructure.s3.get_file_from_s3"
-    )
+    @patch("mcr_meeting.app.infrastructure.s3.get_file_from_s3")
     def test_should_raise_invalid_audio_file_error_when_s3_download_fails(
         self, mock_get_file_from_s3: MagicMock
     ) -> None:
@@ -43,9 +41,7 @@ class TestDownloadAndConcatenateS3AudioChunks:
         )
         assert "S3 connection timeout" in str(exc_info.value)
 
-    @patch(
-        "mcr_meeting.app.infrastructure.s3.get_file_from_s3"
-    )
+    @patch("mcr_meeting.app.infrastructure.s3.get_file_from_s3")
     def test_should_process_successfully_when_valid_chunks(
         self, mock_get_file_from_s3: MagicMock
     ) -> None:
@@ -84,9 +80,7 @@ class TestDownloadAndConcatenateS3AudioChunks:
         assert result.getvalue() == b"audio_data_1audio_data_2audio_data_3"
         assert result.tell() == 0
 
-    @patch(
-        "mcr_meeting.app.infrastructure.s3.get_file_from_s3"
-    )
+    @patch("mcr_meeting.app.infrastructure.s3.get_file_from_s3")
     def test_should_handle_empty_chunks_gracefully(
         self, mock_get_file_from_s3: MagicMock
     ) -> None:
@@ -122,9 +116,7 @@ class TestDownloadAndConcatenateS3AudioChunks:
         assert result.getvalue() == b"valid_data"
         assert result.tell() == 0
 
-    @patch(
-        "mcr_meeting.app.infrastructure.s3.get_file_from_s3"
-    )
+    @patch("mcr_meeting.app.infrastructure.s3.get_file_from_s3")
     def test_should_handle_single_chunk_successfully(
         self, mock_get_file_from_s3: MagicMock
     ) -> None:
@@ -148,9 +140,7 @@ class TestDownloadAndConcatenateS3AudioChunks:
         assert result.getvalue() == b"single_audio_data"
         assert result.tell() == 0
 
-    @patch(
-        "mcr_meeting.app.infrastructure.s3.get_file_from_s3"
-    )
+    @patch("mcr_meeting.app.infrastructure.s3.get_file_from_s3")
     def test_should_handle_large_number_of_chunks(
         self, mock_get_file_from_s3: MagicMock
     ) -> None:
@@ -181,9 +171,7 @@ class TestDownloadAndConcatenateS3AudioChunks:
         assert result.getvalue() == expected_data
         assert result.tell() == 0
 
-    @patch(
-        "mcr_meeting.app.infrastructure.s3.get_file_from_s3"
-    )
+    @patch("mcr_meeting.app.infrastructure.s3.get_file_from_s3")
     def test_should_handle_different_s3_error_types(
         self, mock_get_file_from_s3: MagicMock
     ) -> None:
@@ -218,9 +206,7 @@ class TestDownloadAndConcatenateS3AudioChunks:
                 exc_info.value
             )
 
-    @patch(
-        "mcr_meeting.app.infrastructure.s3.get_file_from_s3"
-    )
+    @patch("mcr_meeting.app.infrastructure.s3.get_file_from_s3")
     def test_should_return_buffer_at_beginning(
         self, mock_get_file_from_s3: MagicMock
     ) -> None:
